+++ app-server/subprojects/bounded_context/place/application/src/main/kotlin/club/staircrusher/place/application/port/in/accessibility/RegisterBuildingAccessibilityUseCase.kt
@@ -13,7 +13,8 @@ import club.staircrusher.stdlib.persistence.TransactionManager
 class RegisterBuildingAccessibilityUseCase(
     private val transactionManager: TransactionManager,
     private val accessibilityApplicationService: AccessibilityApplicationService,
-    private val challengeService: ChallengeService
+    private val challengeService: ChallengeService,
+    private val accessibilityImagePipeline: AccessibilityImagePipeline,
 ) {
     data class RegisterBuildingAccessibilityUseCaseResult(
         val registerBuildingAccessibilityResult: RegisterBuildingAccessibilityResult,
@@ -28,6 +29,9 @@ class RegisterBuildingAccessibilityUseCase(
         val registerResult = accessibilityApplicationService.doRegisterBuildingAccessibility(
             createBuildingAccessibilityParams,
             createBuildingAccessibilityCommentParams
+        )
+        accessibilityImagePipeline.asyncPostProcessImages(
+            registerResult.buildingAccessibility.newEntranceAccessibilityImages + registerResult.buildingAccessibility.newElevatorAccessibilityImages
         )
         val challengeContributions =
             challengeService.contributeToSatisfiedChallenges(
